Bring home orphaned `raster` JsonFormats
